+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 - [Issue #310](https://github.com/influxdb/influxdb/issues/310). Request should support multiple timeseries
+- [Issue #416](https://github.com/influxdb/influxdb/issues/416). Improve the time it takes to drop database
 
 ### Bugfixes